backend/pyproject.toml
[project]
name = "spaceship-backend"
version = "0.1.0"
description = "Spaceship logistics analytics backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.36",
    "psycopg[binary]>=3.2.3",
    "alembic>=1.14.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4.0",  # passlib 1.7.x is incompatible with bcrypt 4.x (AttributeError on __about__)
    "python-multipart>=0.0.12",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "statsmodels>=0.14.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "provider_smoke: hits real LLM providers (opt-in)",
    "eval: NL→tool routing eval suite",
]

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["tests*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
